Updating jenkins file to work with pipeline plugin
Recently in daily tempest integration job, encountered error while
running Jenkinsfile. Issue was related to env.BRANCH_NAME variable
which is not available in pipeline plugin and is used for daily
tempest integration job. This value is available for multibranch
job.

So using GIT_BRANCH env variable instead.
In GIT_BRANCH has similar value as BRANCH_NAME.
In some cases it may have, 'origin/' as suffix so removing it
if its there.