Use all workers for openstack-control-plane
Currently we are using only 2 out of 3 workers (default deployment)
for the control plane deployment of controller pods. This makes the
2 first nodes resource exhausted while the third node does nothing.

Instead, as this is used only for development/CI currently, use
all of the workers for a smooth, no-pod-restart, development/CI
workflow.

This has the side effect of having to separate the labeling of the
compute nodes as the openvswitch label already exists in the
worker-2 so it wont be labeled properly. Its still a good thing
to have as we should label things separatedly in case of errors
and/or retries