Reduce volume size for rabbit
While 5Gb migth be a good size for the default db volumes, for
rabbitmq is way too much.

It doesnt help that airship creates a rabbitmq container per component
so we end up with several 5Gb volumes around for a lot of containers.

Instead set it to a more reasonable size (1Gb) which should allow for
rabbitmq to store to persist to disk with no issues