Deploy and fetch dstat logs for caasp workers
In order to monitor the caasp workers resources when deploying
airship/osh into them we need to somehow gather usage data on
them to analyze down the line.

This PR allows us to export an env var (SOCOK8S_DEPLOY_DSTAT=YES)
that will deploy and launch dstat in the background and a trigger
in the teardown function that will fetch the output data from the
nodes before removing them so we can analyze those further.