Parse subunit output from tempest into junit
Currently we have no way of tracking the output of tempest in
the CI.

This patch adds a new playbook to extract the logs from the tempest
container, transform them into junit files and parse them on CI
for easy tracking of the tempest tests.

What it does its that it creates a temporal container with the tempest
pvc attached in order to extract the data from the tempest run, then
brings it to ansible_host, creates a temporal virtualenv to install
the tools needed for transofrming the data, and then transforms it
into a junit file so the CI can understand the output.

This requires a small change in our tempest deployment, which creates
a workspace at /var/lib/tempest/data, which is where the pvc is mounted,
in order to save the results. This also means that you can run tempest
several times and all the results will be in the same place, accesible
by creating a container and mounting the pvc.