Workaround bad LB issue
Without this patch, helm test would fail, because of kubectl
port-forward tunnel getting closed on more than 50 seconds connection
streams [1].

This is a problem, as helm test is part of our delivery path.

This should fix it by making sure the LB is not used, and we
directly target the master node, until a longer term solution
is found.

[1]: https://bugzilla.suse.com/show_bug.cgi?id=1149211