Disable helm test containers on CI
We are interested in the tempest tests to check for issues.
Running the container tests as part of the CI is useless as we
cannot extract any data from those and thye make the deployment
slower.

Not only that, but those tests are run as part of upstream so they
have already been confirmed to work. Running them again makes no
sense from a CI point of view.