Retry teardown on CI
Currently if one of the steps inside the teardown script fails,
the whole thing will stop. Be it a temporal problem or a real
issue.

This patch wraps the teardown into a retry statement so we try to
cleanup a max of 3 times and then resort to notifying via RocketChat