Use virsh remote by default instead of local socket
Without this patch, terraform by default tries to use the local
socket that is exposed to the container. This is done by default
in the terraform state, which overrides any environment variable.

This is a problem, as we don't expose the whole host /tmp, which
is required for the terraform network creation. In that case,
if the network doesn't exist yet, terraform will fail.

This fixes it by always relying on configuring the virsh
connection string to something non local. In that case, the
network creation succeeds. By making this exposed to the user,
we can simplify the caasp4 script, as it doesn't need the
bind mounts of the hosts for kvm.

Next to this, the terraform example variables are also updated
to fix the instanciation of the caasp cluster for ceph usage
for both kvm and openstack, by adding the required lvm package,
and making sure that the LB ip isn't used.