Add ssh-agent handler
Skuba needs SSH_AUTH_SOCK. In CI we don't need to setup one agent
per workspace, as we are relying on a single key anyway.

This makes it super simple to reuse one single SSH_AUTH_SOCK,
and make sure the identities are still present before doing any
other command. Security wise, it sucks, but so does remote
execution in the same user...
(In other words, the CI configuration doesn't make this worse).