Make auto-deploy tooling a dev tool
Dev could also need to deploy with kubeconfig file.

Without this patch, the dev has to export it manually, or change
its own kubeconfig file (link to home for example), to continue
the deployment.

This makes it easier to use, and reduces the code paths, by making
CI use the dev script.