Wait up to 120s for connections to finish before reloading
Looks like there is a limit of 10 seconds for a connection to
finish when a reload is done on ingress before it gets killed.

This means that long requests like those of openstack can get killed
if there is a backend reload in the start of middle of them.

This bumps it to 120s so we can let those requests finish. This should
not affect anything else as this is only applied to ongoing requests,
so any new requests will get a reloaded backend as this will only
affect the worker dealing with that request