In libvirt deployments, use loadbalancer also as deployer
To not increase the footprint of the libvirt deployment too much. Try
using the loadbalancer node as the osh-deployer node as well.
As the terraform instances don't allow root access via ssh. Make sure
to connect as "sles" and run the ansible tasks for the setting up the
deployer with "become: yes"