Rework tempest deployment
This reworks tempest deployment to be more close to what we do
upstream. Also renames some of the vars to make them more clear.
Tempest network/subnet/pools are now deployed using heat so
its more consistent

Together with this, we change the default values for some of the
vars to be able to have a sane environment with proper values
for vars (i.e. no more docker0)

Also adds a section to add masquerading in our caasp workers so
floating ips work